backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy>=2.0
asyncpg
pydantic
openai
requests
uuid
edge-tts
python-multipart
openai-whisper
FFmpeg